+++ lib/database/repositories/files.repository.ts
@@ -488,7 +488,9 @@ export class FilesRepository extends TaggableBaseRepository<FileEntry> {
       const file = await this.findOneByFilter({ storageKey } as QueryFilter);
 
       if (file) {
+        logger.debug('[Files] File found by storage key', { storageKey });
       } else {
+        logger.debug('[Files] File not found by storage key', { storageKey });
       }
 
       return file;