+++ lib/database/repositories/image-profiles.repository.ts
@@ -44,7 +44,9 @@ export class ImageProfilesRepository extends TaggableBaseRepository<ImageProfile
       } as QueryFilter);
 
       if (profile) {
+        logger.debug('[ImageProfiles] Profile found by name', { userId, name });
       } else {
+        logger.debug('[ImageProfiles] Profile not found by name', { userId, name });
       }
 
       return profile;
@@ -69,7 +71,9 @@ export class ImageProfilesRepository extends TaggableBaseRepository<ImageProfile
       } as QueryFilter);
 
       if (profile) {
+        logger.debug('[ImageProfiles] Default profile found', { userId });
       } else {
+        logger.debug('[ImageProfiles] No default profile found', { userId });
       }
 
       return profile;