+++ lib/database/repositories/plugin-config.repository.ts
@@ -43,7 +43,9 @@ export class PluginConfigRepository extends UserOwnedBaseRepository<PluginConfig
       } as QueryFilter);
 
       if (config) {
+        logger.debug('[PluginConfig] Config found', { userId, pluginName });
       } else {
+        logger.debug('[PluginConfig] Config not found', { userId, pluginName });
       }
 
       return config;